+++ pkg/app/customer/handler/route.go
@@ -197,12 +197,12 @@ func loadFuncMap() {
 			return fmt.Sprintf("%d kB", space)
 		}
 		if space < 1024*1024 {
-			return fmt.Sprintf("%d MB", space/1024)
+			return fmt.Sprintf("%.01f MB", float32(space)/1024)
 		}
 		if space < 1024*1024*1024 {
-			return fmt.Sprintf("%d GB", space/1024/1024)
+			return fmt.Sprintf("%.01f GB", float32(space)/1024/1024)
 		}
-		return fmt.Sprintf("%d TB", space/1024/1024/1024)
+		return fmt.Sprintf("%.02f TB", float32(space)/1024/1024/1024)
 	}
 	// 所有的空间单位必须是int64
 	tplFuncMap["size4Human"] = func(space int64) string {
